Update client. Add support for context. Move things around a little.
