+++ src/App/NetDaemon.App/NetDaemon.App.csproj
@@ -22,7 +22,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="JoySoftware.HassClient" Version="0.5.5-beta" />
+    <PackageReference Include="JoySoftware.HassClient" Version="0.6.0-beta" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.5" />
     <PackageReference Include="System.Reactive" Version="4.4.1" />
   </ItemGroup>
