+++ src/DaemonRunner/DaemonRunner/DaemonRunner.csproj
@@ -21,7 +21,7 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="JoySoftware.HassClient" Version="0.5.5-beta" />
+    <PackageReference Include="JoySoftware.HassClient" Version="0.6.0-beta" />
     <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="3.6.0" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.5" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.5" />
